+++ smarts/core/smarts.py
@@ -211,9 +211,9 @@ class SMARTS(ProviderManager):
         self._last_provider_state = None
         self._reset_agents_only = reset_agents_only  # a.k.a "teleportation"
 
-        # For macOS GUI. See our `BulletClient` docstring for details.
-        # from .utils.bullet import BulletClient
-        # self._bullet_client = BulletClient(pybullet.GUI)
+        # For macOS GUI. See our `BulletClientMACOS` docstring for details.
+        # from .utils.pybullet import BulletClientMACOS
+        # self._bullet_client = BulletClientMACOS(pybullet.GUI)
         self._bullet_client = pybullet.SafeBulletClient(
             pybullet.DIRECT
         )  # pylint: disable=no-member